+++ src/db/models/Role.ts
@@ -27,7 +27,7 @@ Role.init({
     allowNull: false,
     autoIncrement: true,
     primaryKey: true,
-    type: DataTypes.BIGINT
+    type: DataTypes.INTEGER
   },
   roleName: {
     allowNull: true,
